onenand: Factor out read/write/erase/check/dump into onenand_util
In preparation for a OneNAND backend for Fastboot, factor out the OneNAND utilities
