+++ drivers/mtd/onenand/Makefile
@@ -4,7 +4,7 @@
 # Kyungmin Park <kyungmin.park@samsung.com>
 
 ifndef	CONFIG_SPL_BUILD
-obj-$(CONFIG_CMD_ONENAND)	:= onenand_uboot.o onenand_base.o onenand_bbt.o
+obj-$(CONFIG_CMD_ONENAND)	:= onenand_uboot.o onenand_base.o onenand_bbt.o onenand_util.o
 obj-$(CONFIG_SAMSUNG_ONENAND)	+= samsung.o
 else
 obj-y				:= onenand_spl.o
